hot fix paging postService

+++ SocialService.API/Models/DTO/PostDto.cs
@@ -6,7 +6,7 @@
         public int AuthorUserId { get; set; }
         public string? Title { get; set; }
         public string Content { get; set; } = null!;
-        public byte? Rating { get; set; }
+        public double? Rating { get; set; }
         public bool IsReview { get; set; }
         public string Visibility { get; set; } = null!;
         public DateTime CreatedAt { get; set; }
